# docs/swagger.yaml
basePath: /v1
definitions:
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  model.Comment:
    properties:
      content:
        type: string
      reply:
        $ref: '#/definitions/model.User'
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.Emoticon:
    properties:
      category:
        type: string
      id:
        type: integer
      name:
        type: string
      url:
        type: string
    type: object
...
  model.Moment:
    properties:
      list:
        items:
          $ref: '#/definitions/model.MomentList'
        type: array
      user:
        $ref: '#/definitions/model.User'
    type: object
  model.MomentList:
    properties:
      comments:
        items:
          $ref: '#/definitions/model.Comment'
        type: array
      content:
        type: string
      created_at:
        type: integer
      id:
        type: integer
      image:
        type: string
      likes:
        items:
          $ref: '#/definitions/model.User'
        type: array
      location:
        type: string
      type:
        type: integer
      user:
        $ref: '#/definitions/model.User'
      video:
        type: string
    type: object
  model.User:
    properties:
      avatar:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
...
  resource.ApplyListResponse:
    properties:
      status:
        type: integer
      user:
        $ref: '#/definitions/model.User'
    type: object
...
  resource.UserResponse:
    properties:
      area:
        type: string
      avatar:
        type: string
      email:
        type: string
      gender:
        type: integer
      id:
        type: integer
      nickname:
        type: string
      phone:
        type: integer
      sign:
        type: string
      username:
        type: string
    type: object
  user.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/resource.UserResponse'
    type: object
  user.phoneLoginParams:
    properties:
      phone:
        description: 手机号
        example: 13333333333
        type: integer
      verify_code:
        description: 验证码
        example: "888888"
        type: string
    required:
    - phone
    - verify_code
    type: object
...
  user.updateParams:
    properties:
      avatar:
        description: 头像
        example: http://example
        type: string
      nickname:
        description: 昵称
        example: test
        maxLength: 30
        type: string
      sign:
        description: 签名
        example: test
        maxLength: 90
        type: string
    type: object
...
host: http://127.0.0.1:9050
info:
  contact:
    email: test@test.com
    name: test
    url: http://www.swagger.io/support
  license:
    name: MIT
  title: swagger 接口文档
  version: "2.0"
paths:
  /apply/count:
    get:
      description: 待处理申请数量
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":0,"message":"OK","data":1}'
          schema:
            type: string
      summary: 待处理申请数量
      tags:
      - 好友申请
...
  /apply/list:
    get:
      consumes:
      - application/json
      description: 我的申请列表
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 页码
        in: query
        name: p
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/resource.ApplyListResponse'
                  type: array
              type: object
      summary: 我的申请列表
      tags:
      - 好友申请
...
  /emoticon/list:
    get:
      consumes:
      - application/json
      description: 表情包
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 分类
        in: query
        name: cat
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Emoticon'
                  type: array
              type: object
      summary: 表情包
      tags:
      - 表情包
...
  /group/user:
    get:
      consumes:
      - application/json
      description: 获取群成员
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 群ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.User'
                  type: array
              type: object
      summary: 获取群成员
      tags:
      - 群组
...
  /login:
    post:
      consumes:
      - application/json
      description: 通过用户名密码登录
      parameters:
      - description: 用户名
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.authResponse'
              type: object
      summary: 用户登录接口
      tags:
      - 用户
  /login_phone:
    post:
      consumes:
      - application/json
      description: 仅限手机登录
      parameters:
      - description: phone
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.phoneLoginParams'
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/user.authResponse'
              type: object
      summary: 用户登录接口
      tags:
      - 用户
  /logout:
    get:
      consumes:
      - application/json
      description: 用户注销登录
      parameters:
      - description: user token
        in: header
        name: Token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            $ref: '#/definitions/app.Response'
      summary: 用户注销登录
      tags:
      - 用户
...
  /moment/list:
    get:
      consumes:
      - application/json
      description: 动态列表
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: 用户id
        in: query
        name: user_id
        type: integer
      - description: 页码
        in: query
        name: p
        type: integer
      produces:
      - application/json
      responses:
        "0":
          description: 调用成功结构
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Moment'
                  type: array
              type: object
      summary: 动态列表
      tags:
      - 朋友圈
...
  /user/edit:
    post:
      consumes:
      - application/json
      description: Update a user by ID
      parameters:
      - description: 用户令牌
        in: header
        name: Token
        required: true
        type: string
      - description: The user info
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/user.updateParams'
      produces:
      - application/json
      responses:
        "0":
          description: '{"code":0,"msg":"OK","data":{}}'
          schema:
            type: string
      summary: Update a user info by the user identifier
      tags:
      - 用户
...
securityDefinitions:
  Token:
    in: header
    name: Token
    type: apiKey
swagger: "2.0"